refactor: consolidate blocked output handling
